backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
pyyaml==6.0.2
pandas==2.2.3
openpyxl==3.1.5
python-multipart==0.0.12
pytest==8.3.3
httpx==0.27.2
